refactor: View를 입력과 출력 View로 분할
-
